\name{back.Deque}
\alias{back.Deque}
\title{Inspect the Back Element of a Deque}
\description{
This method returns the back element of a \code{Deque}.
}
\details{
\code{aDeque$back()}
}
\value{
  \code{back} returns the back element of the \code{Deque} in question, or \code{NULL} if the \code{Deque} is empty.
}
\author{John Hughes}
\keyword{methods}
