\name{capacity.Heap}
\alias{capacity.Heap}
\title{Determine a Heap's Memory Allocation}
\description{
This method returns the of elements the \code{Heap} is capable of holding. The method can be used in conjunction with \code{size} and \code{contract} to ensure optimum memory usage.
}
\details{
\code{aHeap$capacity()}
}
\value{
  \code{capacity} returns a nonnegative integer.
}
\author{John Hughes}
\keyword{methods}
