\name{clone.Container}
\alias{clone.Container}
\title{Clone a Container}
\description{
This method makes a deep copy of a given \code{Container}.
}
\details{
\code{copy = aStack$clone()}
}
\value{
  \code{clone} returns a deep copy of the \code{Container} in question.
}
\author{John Hughes}
\keyword{methods}
