\name{Container}
\alias{Container}
\title{Constructor for Class Container}
\description{
This method instanciates class \code{Container}, but this method should not be called directly, for \code{Container} is an abstract class. See the package vignette for details.
}
\author{John Hughes}
\keyword{package}
\keyword{methods}
