\name{current.Iterator}
\alias{current.Iterator}
\title{Inspect an Iterator's Current Element of Interest}
\description{
This method returns the current element of the container to which the \code{Iterator} is attached, if possible.
}
\details{
\code{e = it$current()}
}
\value{
  \code{current} returns the element over which the \code{Iterator} is currently positioned, or \code{NULL} if the \code{Iterator} is invalid.
}
\author{John Hughes}
\keyword{methods}
