\name{find.BST}
\alias{find.BST}
\title{Search a Binary Search Tree}
\description{
This method searches a \code{BST}.
}
\details{
\code{e = aBST$find(target)}
}
\value{
  \code{find} returns the found element, or \code{NULL} if the target element is absent.
}
\author{John Hughes}
\keyword{methods}
