\name{front.Queue}
\alias{front.Queue}
\title{Inspect the Front Element of a Queue}
\description{
This method returns the front element of a \code{Queue}.
}
\details{
\code{aQueue$front()}
}
\value{
  \code{front} returns the front element of the \code{Queue} in question, or \code{NULL} if the \code{Queue} is empty.
}
\author{John Hughes}
\keyword{methods}
