\name{Heap}
\alias{Heap}
\title{Constructor for Class Heap}
\description{
This method instanciates class \code{Heap}, but this method should not be called directly, for \code{Heap} is an abstract class. See the package vignette for details.
}
\author{John Hughes}
\keyword{package}
\keyword{methods}
