\name{height.BST}
\alias{height.BST}
\title{Determine the Height of a Binary Search Tree}
\description{
This method returns the height of a \code{BST}.
}
\details{
\code{ht = aBST$height()}
}
\value{
  \code{height} returns -1 if the tree is empty. Otherwise, the returned value is the depth of the \code{BST}'s deepest node.
}
\author{John Hughes}
\keyword{methods}
