\name{isEmpty.Container}
\alias{isEmpty.Container}
\title{Determine Whether a Container is Empty}
\description{
This method returns \code{TRUE} (\code{FALSE}) for an empty (nonempty) \code{Container}.
}
\details{
\code{empty = aContainer$isEmpty()}
}
\value{
  \code{isEmpty} returns a logical value.
}
\author{John Hughes}
\keyword{methods}
