\name{isItem.Iterator}
\alias{isItem.Iterator}
\title{Determine Whether an Iterator is Valid}
\description{
This method returns \code{TRUE} (\code{FALSE}) for a valid (invalid) \code{Iterator}.
}
\details{
\code{valid = it$isItem()}
}
\value{
  \code{isItem} returns a logical value.
}
\author{John Hughes}
\keyword{methods}
