\name{iterator.Heap}
\alias{iterator.Heap}
\title{Get a Heap Iterator}
\description{
This method returns an \code{Iterator} for the \code{Heap} in question. The iterator starts at the root and traverses a given level of the tree from left to right before descending to the next level.
}
\details{
\code{it = aHeap$iterator()}
}
\value{
  \code{iterator} returns an instance of \code{Iterator}.
}
\author{John Hughes}
\keyword{methods}
