\name{iterator.Queue}
\alias{iterator.Queue}
\title{Get a Queue Iterator}
\description{
This method returns an \code{Iterator} for the \code{Queue} in question. Both forward and reverse iteration are supported.
}
\details{
\code{it = aQueue$iterator()}
}
\value{
  \code{iterator} returns an instance of \code{Iterator}.
}
\author{John Hughes}
\keyword{methods}
