\name{iterator.Stack}
\alias{iterator.Stack}
\title{Get a Stack Iterator}
\description{
This method returns an \code{Iterator} for the \code{Stack} in question. The iterator traverses the \code{Stack} from top to bottom.
}
\details{
\code{it = aStack$iterator()}
}
\value{
  \code{iterator} returns an instance of \code{Iterator}.
}
\author{John Hughes}
\keyword{methods}
