\name{maximum.BST}
\alias{maximum.BST}
\title{Inspect the Largest Element of a Binary Search Tree}
\description{
This method returns the largest element of a \code{BST}.
}
\details{
\code{largest = aBST$maximum()}
}
\value{
  \code{maximum} returns the largest element of the \code{BST} in question, or \code{NULL} if the \code{BST} is empty.
}
\author{John Hughes}
\keyword{methods}
