\name{minimum.BST}
\alias{minimum.BST}
\title{Inspect the Smallest Element of a Binary Search Tree}
\description{
This method returns the smallest element of a \code{BST}.
}
\details{
\code{smallest = aBST$minimum()}
}
\value{
  \code{minimum} returns the smallest element of the \code{BST} in question, or \code{NULL} if the \code{BST} is empty.
}
\author{John Hughes}
\keyword{methods}
