\name{moveBackward.Iterator}
\alias{moveBackward.Iterator}
\title{Move an Iterator Backward}
\description{
This method moves an \code{Iterator} backward by one element, provided that reverse iteration is supported for the kind of container to which the iterator is attached.
}
\details{
\code{it$moveBackward()}
}
\author{John Hughes}
\keyword{methods}
