\name{peekMax.MaxHeap}
\alias{peekMax.MaxHeap}
\title{Inspect the Root of a Max-Heap}
\description{
This method returns the root element of a \code{MaxHeap}.
}
\details{
\code{largest = aMaxHeap$peekMax()}
}
\value{
  \code{peekMax} returns the root element of the \code{MaxHeap} in question, or \code{NULL} if the \code{MaxHeap} is empty.
}
\author{John Hughes}
\keyword{methods}
