\name{peekMin.MinHeap}
\alias{peekMin.MinHeap}
\title{Inspect the Root of a Min-Heap}
\description{
This method returns the root element of a \code{MinHeap}.
}
\details{
\code{smallest = aMinHeap$peekMin()}
}
\value{
  \code{peekMin} returns the root element of the \code{MinHeap} in question, or \code{NULL} if the \code{MinHeap} is empty.
}
\author{John Hughes}
\keyword{methods}
