\name{pop.Stack}
\alias{pop.Stack}
\title{Remove the Top Element of a Stack}
\description{
This method removes and returns the top element of a \code{Stack}.
}
\details{
\code{e = aStack$pop()}
}
\value{
  \code{pop} removes and returns the top element of the \code{Stack} in question, or returns \code{NULL} if the \code{Stack} is empty.
}
\author{John Hughes}
\keyword{methods}
