\name{predecessor.BST}
\alias{predecessor.BST}
\title{Get the In-Order Predecessor of an Element of a Binary Search Tree}
\description{
This method returns the in-order predecessor of the target element.
}
\details{
\code{before = aBST$predecessor(target)}
}
\value{
  \code{predecessor} returns the in-order predecessor of the target element of the \code{BST}, or \code{NULL} if the target has no predecessor.
}
\author{John Hughes}
\keyword{methods}
