\name{removeBack.Deque}
\alias{removeBack.Deque}
\title{Remove the Back Element of a Deque}
\description{
This method removes and returns the back element of a \code{Deque}.
}
\details{
\code{e = aDeque$removeBack()}
}
\value{
  \code{removeBack} removes and returns the back element of the \code{Deque} in question, or returns \code{NULL} if the \code{Deque} is empty.
}
\author{John Hughes}
\keyword{methods}
