\name{removeFront.Queue}
\alias{removeFront.Queue}
\title{Remove the Front Element of a Queue}
\description{
This method removes and returns the front element of a \code{Queue}.
}
\details{
\code{e = aQueue$removeFront()}
}
\value{
  \code{removeFront} removes and returns the front element of the \code{Queue} in question, or returns \code{NULL} if the \code{Queue} is empty.
}
\author{John Hughes}
\keyword{methods}
