\name{removeMax.MaxHeap}
\alias{removeMax.MaxHeap}
\title{Remove the Root Element of a Max-Heap}
\description{
This method removes and returns the largest element of a \code{MaxHeap}.
}
\details{
\code{largest = aMaxHeap$removeMax()}
}
\value{
  \code{removeMax} removes and returns the root element of the \code{MaxHeap} in question, or returns \code{NULL} if the \code{MaxHeap} is empty.
}
\author{John Hughes}
\keyword{methods}
