\name{removeMin.MinHeap}
\alias{removeMin.MinHeap}
\title{Remove the Root Element of a Min-Heap}
\description{
This method removes and returns the smallest element of a \code{MinHeap}.
}
\details{
\code{smallest = aMinHeap$removeMin()}
}
\value{
  \code{removeMin} removes and returns the root element of the \code{MinHeap} in question, or returns \code{NULL} if the \code{MinHeap} is empty.
}
\author{John Hughes}
\keyword{methods}
