\name{rstart.Iterator}
\alias{rstart.Iterator}
\title{Prepare an Iterator for Reverse Iteration}
\description{
This method positions the \code{Iterator} over the back element of a \code{Queue} or a \code{Deque}. If this method is called on a \code{Stack} or \code{Heap} iterator, said iterator becomes invalid.
}
\details{
\code{it$rstart()}
}
\author{John Hughes}
\keyword{methods}
