\name{size.Container}
\alias{size.Container}
\title{Determine the Size of a Container}
\description{
This method returns the number of elements in a given \code{Container}.
}
\details{
\code{sz = aContainer$size()}
}
\value{
  \code{size} returns a nonnegative integer.
}
\author{John Hughes}
\keyword{methods}
