\name{SplayTree}
\alias{SplayTree}
\title{Constructor for Class SplayTree}
\description{
This method instanciates class \code{SplayTree}, which specializes \code{BST}. See the package vignette for details.
}
\usage{
aSplayTree = SplayTree()
}
\author{John Hughes}
\keyword{package}
\keyword{methods}
