\name{start.Iterator}
\alias{start.Iterator}
\title{Prepare an Iterator for Forward Iteration}
\description{
This method positions the \code{Iterator} over the top (front, front, root) element of a \code{Stack} (\code{Queue}, \code{Deque}, \code{Heap}).
}
\details{
\code{it$start()}
}
\author{John Hughes}
\keyword{methods}}
