\name{successor.BST}
\alias{successor.BST}
\title{Get the In-Order Successor of an Element of a Binary Search Tree}
\description{
This method returns the in-order successor of the target element.
}
\details{
\code{after = aBST$successor(target)}
}
\value{
  \code{successor} returns the in-order successor of the target element of the \code{BST}, or \code{NULL} if the target has no successor.
}
\author{John Hughes}
\keyword{methods}
