\name{top.Stack}
\alias{top.Stack}
\title{Inspect the Top Element of a Stack}
\description{
This method returns the top element of a \code{Stack}.
}
\details{
\code{e = aStack$top()}
}
\value{
  \code{top} returns the top element of the \code{Stack} in question, or \code{NULL} if the \code{Stack} is empty.
}
\author{John Hughes}
\keyword{methods}
